Prevent context issue w/ django.contrib.messages
Due to the situation we have at work, I encounter a problem whereby the
variable 'messages' in the template render context is overwritten by the
variable 'messages' provided by django.contrib.messages which is passed
in via a RequestContext.

This commit simply renames 'messages' to 'rosetta_messages' to avoid
this problem.